Design : 마이페이지 스타일링 수정
