+++ package.json
@@ -21,7 +21,7 @@
     "react-dom": "^18",
     "react-intersection-observer": "^9.13.0",
     "react-kakao-maps-sdk": "^1.1.27",
-    "swiper": "^11.1.5",
+    "swiper": "^11.1.8",
     "zustand": "^4.5.4"
   },
   "devDependencies": {
