+++ src/components/mypage/profile/ProfileCarousel.tsx
@@ -3,6 +3,7 @@ import { Swiper, SwiperSlide } from "swiper/react";
 
 import "swiper/css";
 import "swiper/css/pagination";
+import "@/app/globals.css";
 
 import { Pagination, Autoplay } from "swiper/modules";
 import Image from "next/image";
@@ -12,9 +13,6 @@ const ProfileCarousel: React.FC = () => {
 
   return (
     <Swiper
-      pagination={{
-        dynamicBullets: true,
-      }}
       modules={[Pagination, Autoplay]}
       className="mySwiper h-full rounded-xl"
       style={{ width: "100%", height: "100%" }}
