+++ src/components/mypage/wishlist/MyPageCarousel.module.css
@@ -1,7 +1,3 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
 .my-page .swiper-pagination {
   bottom: 20px !important;
 }
